fix: bw cli was not c orrectly installed

+++ install.sh
@@ -4,16 +4,17 @@ set -e
 
 sudo dnf install -y unzip curl chezmoi
 
-mkdir -p ~/.local/bin
-curl -Lso bw.zip "https://vault.bitwarden.com/download/?app=cli&platform=linux"
-unzip -q bw.zip -d ~/.local/bin/
-chmod +x ~/.local/bin/bw
-rm bw.zip
+if ! command -v bw &> /dev/null; then
+    curl -Lso bw.zip "https://vault.bitwarden.com/download/?app=cli&platform=linux"
+    sudo unzip -q bw.zip -d /usr/local/bin/
+    sudo chmod +x /usr/local/bin/bw
+    rm bw.zip
+fi
 
-export PATH="$HOME/.local/bin:$PATH"
-
-bw config server https://vault.bitwarden.eu
-bw login
+if bw status | grep -q '"status":"unauthenticated"'; then
+    bw config server https://vault.bitwarden.eu
+    bw login
+fi
 
 export BW_SESSION=$(bw unlock --raw)
